BITZ/server/requirements.txt
openai
gradio
folium
httpx==0.27.2
plotly
langchain
langchain-openai
flask
flask-cors
python-dotenv
markdown
google-cloud-storage
gunicorn
pymongo